+++ libft/srcs/maths/ft_recursive_power.c
@@ -12,7 +12,7 @@
 
 #include "../../includes/libft.h"
 
-int		ft_recursive_power(int nb, int power)
+long		ft_recursive_power(long nb, int power)
 {
 	int nb_init;
